requirements.txt
requests
bs4
langchain
openai
python-dotenv
chromadb
pickle
tiktoken